django/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-bootstrap-v5==1.0.11
django-bootstrap5==23.1
django-filter==23.2
djangorestframework==3.14.0
docker==6.1.1
idna==3.4
kombu==5.2.4
Markdown==3.4.3
packaging==23.1
prompt-toolkit==3.0.38
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.5.1
